requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

langchain-groq>=0.1.9

sentence-transformers>=3.0.0

faiss-cpu>=1.7.4

requests>=2.31.0
beautifulsoup4>=4.12.0

python-dotenv>=1.0.0

streamlit>=1.35.0